crypto: rsa: define RsaKey key interface
Define the new struct RsaKey for providing all RSA key related
functionality through a single interface. It stores a pair of a public
key alongside with an optional associated private key and provides
primitives to convert to/from TPM interface types as well as the
basic key operations: generate, encrypt and  decrypt.

Signed-off-by: Nicolai Stange <nstange@suse.de>