crypto: ecc: implement random scalar generation
The generation of random scalars in the range from 1 to the curve order
(exclusive) is needed for several ECC related primitives: key generation
to begin with, but also for e.g. the randomized signature schemes.

Implement the generation of random scalars in said range by the method
of oversampling, in accordance with TCG TPM2 Library, Part 1, section
C.5 ("ECC Key Generation").

Signed-off-by: Nicolai Stange <nstange@suse.de>